+++ app/models/concerns/admin_column.rb
@@ -5,13 +5,21 @@ module AdminColumn
   included do
     class << self
       attr_reader :non_admin_columns
+
+      def additional_admin_columns
+        columns.select { |c| @additional_admin_columns.include?(c.name) }
+      end
     end
 
     @non_admin_columns = []
+    @additional_admin_columns = []
   end
 
   def for_admin_column
-    columns = translated_columns + self.class.content_columns
+    columns = translated_columns +
+              self.class.content_columns +
+              self.class.additional_admin_columns
+
 
     reject_non_admin_columns(columns).each do |column|
       yield(column.name.humanize,
@@ -36,4 +44,5 @@ module AdminColumn
       []
     end
   end
+
 end