+++ doc/CHANGES.md
@@ -2,6 +2,14 @@
 
 ## Highlighted Features
 
+* Log an `InfoRequestEvent` when updating response handling attributes in
+  `InfoRequest.stop_new_responses_on_old_requests` (Gareth Rees)
+* Show that a request is part of a batch on the request page in the admin
+  interface (Gareth Rees)
+* Add "Rejected incoming count" do the request page in the admin interface
+  (Gareth Rees)
+* Highlight non-default states of "Allow new responses from" in the admin
+  interface (Gareth Rees)
 * Add collapse/expand to request correspondence (Zarino Zappia)
 * Fix downloading a Zip of entire request when the request contains a resent
   message (Gareth Rees)
@@ -16,6 +24,9 @@
   `YOUR_THEME_ROOT/assets/images`, to be shown next to pages from your site when
   shared on Facebook. You can just duplicate `logo-opengraph.png` if you don't
   have specific Pro branding.
+* We've removed the spring preloader as it didn't seem to provide much benefit.
+* `InfoRequest.get_last_event` is deprecated and will be removed in 0.33. Please
+  use `InfoRequest.last_event`.
 
 ### Changed Templates